requirements.txt
fusepy==3.0.1
boto3==1.34.34
watchdog==3.0.0
click==8.1.7
cryptography==41.0.8
python-dateutil==2.8.2
requests==2.31.0
pyyaml==6.0.1
colorlog==6.8.0 